backend/data_pipeline/requirements.txt
qdrant-client
google-genai
